Added hard links to bash_rc, aliases and vimrc
Bash_RC (and aliases) now automatically include all shell scripts in the
ShellScrip directory, script that include

* AddMIT
Add an MIT license file to the cwd

* BuildMarkdown
Takes all in ./md/*.md and converts to ./*.html, styled after github md

* _colors, _globals
Global shell color codes and variables, used via
    echo -e "${COL_Green}this${COL_NC}"

* CheatSheet.sh
Prints out (in terminal) usage for the shell scripts, and explanations
for aliased commands and functions in the bash

* cheatsheets/
A bunch of md files for almost all the languages. May use later, for
in-terminal / in-vim lookup tables (although already have zeam, so...)

-----------------

TODO:

* Build out cheatsheets lookup or strip out.
* Build out CheatSheet.sh to a full help, with interactive menus
* Add a command builder, because who the hell wants to remember all the
SED and GREP commands on their own?